feat: Review 기능 작성
feat: Review 기능 작성
feat: Review, ReviewImage Entity @Setter 어노테이션 추가
feat: Qna 기능 수정

+++ cosmeticShop/src/main/java/Midas/cosmeticShop/service/QnaService.java
@@ -2,10 +2,13 @@ package Midas.cosmeticShop.service;
 
 import Midas.cosmeticShop.dto.QnaDTO;
 import Midas.cosmeticShop.dto.QnaListDTO;
+import Midas.cosmeticShop.dto.QnaPostDTO;
 import Midas.cosmeticShop.entity.Qna;
 import Midas.cosmeticShop.entity.Users.User;
 import Midas.cosmeticShop.repository.QnaRepository;
 import Midas.cosmeticShop.repository.Users.UserRepository;
+import jakarta.persistence.EntityNotFoundException;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 
@@ -46,40 +49,44 @@ public class QnaService {
 
     public QnaDTO getQnaDetail (Long qnaId) {
         Qna qna = QnaRepo.findById(qnaId)
-                .orElseThrow();
+                .orElseThrow(() -> new EntityNotFoundException("QnA가 존재하지 않습니다."));
         return new QnaDTO(qna);
     }
 
-    public void postQna (String questionTitle, String content,String userId) {
+    public void postQna (QnaPostDTO qnaPostDTO,String userId) {
         Qna qna = new Qna();
         User user = UserRepo.findByUserId(userId)
-                        .orElseThrow();
+                        .orElseThrow(() -> new EntityNotFoundException("사용자가 존재하지 않습니다."));
         qna.setUser(user);
-        qna.setQuestionTitle(questionTitle);
-        qna.setContent(content);
+        qna.setQuestionTitle(qnaPostDTO.getQuestionTitle());
+        qna.setContent(qnaPostDTO.getContent());
         qna.setQuestionedAt(LocalDateTime.now());
         QnaRepo.save(qna);
     }
 
-    public void putQna (Long qnaId, String questionTitle, String content) {
+    public void putQna (Long qnaId, QnaPostDTO qnaPostDTO, String userId) {
         Qna qna = QnaRepo.findById(qnaId)
-                .orElseThrow();
-        qna.setQuestionTitle(questionTitle);
-        qna.setContent(content);
+                .orElseThrow(() -> new EntityNotFoundException("QnA가 존재하지 않습니다."));
+        if(!qna.getUser().getUserId().equals(userId)) //작성자가 맞는지 확인
+            throw new AccessDeniedException("본인이 작성한 QnA만 수정할 수 있습니다.");
+        qna.setQuestionTitle(qnaPostDTO.getQuestionTitle());
+        qna.setContent(qnaPostDTO.getContent());
         QnaRepo.save(qna);
     }
 
     public void putQnaAnswer (Long qnaId, String answer) {
         Qna qna = QnaRepo.findById(qnaId)
-                .orElseThrow();
+                .orElseThrow(() -> new EntityNotFoundException("QnA가 존재하지 않습니다."));
         qna.setAnswer(answer);
         qna.setAnsweredAt(LocalDateTime.now());
         QnaRepo.save(qna);
     }
 
-    public  void deleteQna (Long qnaId) {
+    public  void deleteQna (Long qnaId, String userId) {
         Qna qna = QnaRepo.findById(qnaId)
-                .orElseThrow();
+                .orElseThrow(() -> new EntityNotFoundException("QnA가 존재하지 않습니다."));
+        if(!qna.getUser().getUserId().equals(userId))
+            throw new AccessDeniedException("본인이 작성한 QnA만 삭제할 수 있습니다.");
         QnaRepo.deleteById(qnaId);
     }
